pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vibewall"
version = "0.1.0"
description = "Hallucination firewall for AI coding agents"
requires-python = ">=3.12"
dependencies = [
    "mitmproxy>=10",
    "aiohttp",
    "python-whois",
    "rapidfuzz",
    "click",
    "structlog",
]

[project.scripts]
vibewall = "vibewall.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/vibewall"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-httpserver",
    "aioresponses",
]
